refactor [#35] header 관련 파일을 폴더로 묶음
