+++ src/features/rolling-paper/components/header/rolling-paper-header.jsx
@@ -1,17 +1,17 @@
 import EmojiPicker from "emoji-picker-react";
 import styled from "styled-components";
-import addImage from "../../../assets/ic-face-smile-add.svg";
-import shareImage from "../../../assets/ic-share.svg";
-import { OutlinedButton } from "../../../components/button/button";
-import BUTTON_SIZE from "../../../components/button/button-size";
-import Colors from "../../../components/color/colors";
-import Popover from "../../../components/popover/popover";
-import POPOVER_ALIGNMENT from "../../../components/popover/popover-alignment";
-import Toast from "../../../components/toast/toast";
-import { useMedia } from "../../../hooks/use-media";
-import { useToast } from "../../../hooks/use-toast";
-import { shareRollingPaper } from "../../../libs/kakao/kakao-service";
-import { media } from "../../../utils/media";
+import addImage from "../../../../assets/ic-face-smile-add.svg";
+import shareImage from "../../../../assets/ic-share.svg";
+import { OutlinedButton } from "../../../../components/button/button";
+import BUTTON_SIZE from "../../../../components/button/button-size";
+import Colors from "../../../../components/color/colors";
+import Popover from "../../../../components/popover/popover";
+import POPOVER_ALIGNMENT from "../../../../components/popover/popover-alignment";
+import Toast from "../../../../components/toast/toast";
+import { useMedia } from "../../../../hooks/use-media";
+import { useToast } from "../../../../hooks/use-toast";
+import { shareRollingPaper } from "../../../../libs/kakao/kakao-service";
+import { media } from "../../../../utils/media";
 import RollingPaperReactions from "./rolling-paper-reactions";
 import RollingPaperSenders from "./rolling-paper-senders";
 import RollingPaperSharePopover from "./rolling-paper-share-popover";
